factor out file checking
